--- backend/src/Schema/Auth/UserSessionSchema.py
@@ -1,12 +1,0 @@
-from datetime import datetime
-from uuid import UUID
-from pydantic import BaseModel
-
-from src.Schema.User.UserResponseSchema import UserResponseSchema
-
-class UserSessionListSchema(BaseModel):
-    user: UserResponseSchema
-    id: UUID
-    ip: str
-    validUntil: datetime
-    timestamp: datetime
