+++ backend/src/Utils/env.py
@@ -2,7 +2,7 @@ from dotenv import load_dotenv
 import os
 
 if not load_dotenv("./.env"):
-    print("Erro ao carregar o arquivo \".env\". Verifique se ele existe e está no diretório correto ou se contem alguma chave.")
+    raise RuntimeError("Erro ao carregar o arquivo \".env\". Verifique se ele existe e está no diretório correto ou se contem alguma chave.")
 
 def get_env_var(var_name: str) -> str:
     key = os.environ.get(var_name, None)
