+++ frontend/package.json
@@ -13,13 +13,9 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.2.1",
-    "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-dialog": "^1.1.15",
-    "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-navigation-menu": "^1.2.14",
-    "@radix-ui/react-select": "^2.2.6",
     "@radix-ui/react-separator": "^1.1.7",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tooltip": "^1.2.8",
@@ -27,15 +23,11 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.540.0",
-    "motion": "^12.23.12",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "react-hook-form": "^7.62.0",
     "react-router-dom": "^7.8.1",
     "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.1.12",
-    "tailwindcss-animate": "^1.0.7",
-    "zod": "^4.1.5"
+    "tailwindcss": "^4.1.12"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
@@ -50,10 +42,5 @@
     "globals": "^16.3.0",
     "tw-animate-css": "^1.3.7",
     "vite": "^7.1.2"
-  },
-  "engines": {
-    "node": "22.x",
-    "npm": "10.9.x"
   }
 }
-
