+++ frontend/src/components/footer/index.jsx
