+++ frontend/src/pages/Register/index.css
@@ -1,0 +1,18 @@
+﻿.cadastro-page {
+  min-height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 3rem;
+}
+.cadastro.card {
+  width: 520px;
+  max-width: 95%;
+}
+
+.cadastro form input {
+  margin-bottom: 0.5rem;
+}
+.cadastro h1 {
+  font-size: 1.6rem;
+}
